requirements.txt
playwright
python-dotenv
pandas
gspread
google-auth
requests
